refactor : common폴더 components 하위로 이동
